LavaHack b0.1.6.5-8 | Changes #84: Client:     Modules:         Added NameTags -> NoBots         Moved CameraClip from RENDER to PLAYER         Moved Weather from RENDER to MISC         Removed Spin         Removed Reverse         Removed WeaknessLog         Removed old alpha fade from AutoRer
    Guis:
        Improved HalqGui

    Stuff:
        Added Depth feature to RenderingRewritePattern(WIP)
        Added new alpha fade to SlideRenderingRewritePattern
        Updated Rendering(minor changes)
        Improved CrystalUtils::calculateDamage a bit
        Fixed Scale feature of RenderingRewritePattern

        SlideRenderingRewritePattern -> AlphaFade:
            Added Moving -> Lengths -> AlphaFade
            Added Moving -> Easings -> AlphaFade

+++ src/main/java/com/kisman/cc/gui/halq/components/sub/colorpicker/slider/sliders/AlphaSlider.kt
@@ -60,7 +60,7 @@ class AlphaSlider(
             squareIndex += checkerBoardSquareSize
         }
 
-        Render2DUtil.drawLeftGradientRect(x, y_, x + width_, y_ + height, color?.withAlpha(1)?.rgb!!, 0)
+        Render2DUtil.drawLeftGradientRect(x.toDouble(), y_.toDouble(), x.toDouble() + width_, y_.toDouble() + height, color?.withAlpha(1)?.rgb!!, 0)
         val sliderMinX: Int = (x + width_ - width_ * color?.a1!!).toInt()
         Render2DUtil.drawRect(sliderMinX - 1, y_, sliderMinX + 1, y_ + height, -1)
